+++ src/ServiceProvider.php
@@ -4,7 +4,7 @@ namespace DreamFactory\Core\Salesforce;
 use DreamFactory\Core\Components\ServiceDocBuilder;
 use DreamFactory\Core\Enums\ServiceTypeGroups;
 use DreamFactory\Core\Salesforce\Models\SalesforceConfig;
-use DreamFactory\Core\Salesforce\Services\SalesforceDb;
+use DreamFactory\Core\Salesforce\Services\Salesforce;
 use DreamFactory\Core\Services\ServiceManager;
 use DreamFactory\Core\Services\ServiceType;
 
@@ -20,14 +20,14 @@ class ServiceProvider extends \Illuminate\Support\ServiceProvider
                 new ServiceType([
                     'name'            => 'salesforce_db',
                     'label'           => 'Salesforce',
-                    'description'     => 'Database service for Salesforce connections.',
-                    'group'           => ServiceTypeGroups::DATABASE,
+                    'description'     => 'Database service with SOAP and/or OAuth authentication support for Salesforce connections.',
+                    'group'           => [ServiceTypeGroups::DATABASE, ServiceTypeGroups::OAUTH],
                     'config_handler'  => SalesforceConfig::class,
                     'default_api_doc' => function ($service) {
-                        return $this->buildServiceDoc($service->id, SalesforceDb::getApiDocInfo($service));
+                        return $this->buildServiceDoc($service->id, Salesforce::getApiDocInfo($service));
                     },
                     'factory'         => function ($config) {
-                        return new SalesforceDb($config);
+                        return new Salesforce($config);
                     },
                 ])
             );